refactor: change expenses folder structure
The idea is separate the test from the source code in tha package and
define the following standard

expenses.models: Contains all types that represent an expense
expenses.repository: represents data access layers
expenses.service:  represents the business layer
expense.controller: represents the presenter related to RestFul endpoint
expense.resolver: represents the presenter related to Graphql.
