+++ packages/domains/expense/package.json
@@ -3,5 +3,12 @@
     "version": "0.0.0",
     "description": "Expenses domain logic",
     "private": true,
-    "dependencies": {}
+    "dependencies": {
+      "@nc/utils": "^0.0.0",
+      "express": "^4.17.0"
+    },
+    "devDependencies": {
+      "@nc/test-utils": "0.0.1"
+      
+    }
   }
