--- packages/domains/expense/status.ts
@@ -1,4 +1,0 @@
-export enum Status {
-  Pending = 'pending',
-  Processed = 'Processed'
-}
